+++ src/tightdb/objc/TDBQuery_noinst.h
@@ -20,6 +20,9 @@
 
 #import <Foundation/Foundation.h>
 
+#include <tightdb/table_view.hpp>
+
 @interface TDBQuery (noinst)
 -(tightdb::Query&)getNativeQuery;
+-(void)setTableView:(tightdb::TableView&)tableView;
 @end